pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = "Adaptive learning system powered by LangGraph"
requires-python = ">=3.12"
dependencies = [
    "langgraph>=0.4.0",
    "langchain>=0.3.0",
    "langchain-huggingface>=0.3.0",
    "python-dotenv>=1.0.0",
    "json-repair>=0.30.0",
    "pydantic[email]>=2.0",
    "pydantic-settings>=2.0",
    "rich>=13.0",
    "fastapi>=0.110.0",
    "uvicorn>=0.29.0",
    "sqlalchemy>=2.0.0",
    "passlib[bcrypt]>=1.7.4",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.9",
    "bcrypt>=4.1.2",
    "langgraph-checkpoint-sqlite>=2.0.0",
    "asyncpg>=0.29.0",
    "psycopg[binary]>=3.1.0",
    "langgraph-checkpoint-postgres>=2.0.0",
    "redis>=5.0.0",
    "arq>=0.25.0",
    "alembic>=1.13.0",
    "tavily-python>=0.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]

[project.scripts]
agent = "agent.cli:main"

[tool.setuptools.packages.find]
include = ["agent*"]

[build-system]
requires = ["setuptools>=70.0"]
build-backend = "setuptools.build_meta"
